+++ src/theme/components/template.ts
@@ -1,81 +1,69 @@
-// From @mixin breakout:
-// Used for elements that should be edge-to-edge.
+/**
+ * Grid layout based on https://www.joshwcomeau.com/css/full-bleed/
+ */
+
 const breakout = {
-  marginEnd: "calc(-50vw + 50%)",
-  marginStart: "calc(-50vw + 50%)",
-};
-// From the `wrapper` SCSS mixin.
-// TODO: Eventually we may be able to put shared JS style objects
-// into a utils file for other mixins that are shared.
-const wrapperStyles = {
-  marginY: 0,
-  marginX: "auto",
-  maxWidth: "1280px",
-  paddingTop: 0,
-  paddingBottom: 0,
-  paddingEnd: "s",
-  paddingStart: "s",
   width: "100%",
+  // This could be "1 / 4" and it would mean the same. This is
+  // "future-proof" the grid column assignment to the last column.
+  gridColumn: "1 / -1",
 };
 
 const Template = {
   baseStyle: {
     boxSizing: "border-box",
     color: "ui.black",
-    // Users with non-overlay scrollbars have a small horizontal scrollbar from
-    // our breakout mixin. Fix identified here: https://cloudfour.com/thinks/breaking-out-with-viewport-units-and-calc/#one-big-dumb-caveat
-    overflowX: "hidden",
-    "*": { boxSizing: "inherit" },
-    "*::after": { boxSizing: "inherit" },
-    "*::before": { boxSizing: "inherit" },
+    display: "grid",
+    gridTemplateColumns: `
+      1fr
+      min(1280px, 100%)
+      1fr`,
+    // Set all elements to start on the second 1280px grid column.
+    "> *": {
+      gridColumn: "2",
+    },
   },
   sizes: {},
   defaultProps: {},
 };
-const TemplateHeader = {
-  baseStyle: {
-    marginBottom: "page.hstack", // --nypl-space-l
-  },
-};
+// Elements that need to breakout will span outside
+// the center 1280px grid column.
 const TemplateBreakout = {
-  baseStyle: {
-    ...breakout,
-    width: "100vw",
-    marginBottom: "page.hstack", // --nypl-space-l
-  },
+  baseStyle: breakout,
 };
 const TemplateContent = {
   baseStyle: {
-    ...wrapperStyles,
-    display: "block",
-    marginBottom: "page.hstack", // --nypl-space-l
+    display: "grid",
+    gridTemplateColumns: "1fr",
+    marginY: "page.hstack",
+    paddingY: 0,
+    paddingX: "s",
   },
+  // With left or right sidebars, we need to set two grid columns and
+  // the column for the sidebar is max 288px width.
   variants: {
-    sidebar: {
-      display: "flex",
-      flexFlow: {
-        base: "column nowrap",
-        md: "row wrap",
-      },
+    left: {
+      gridTemplateColumns: { md: "288px 1fr" },
+    },
+    right: {
+      gridTemplateColumns: { md: "1fr 288px" },
     },
   },
 };
 const TemplateContentTop = {
   baseStyle: {
-    flex: "1 0 100%",
-    marginBottom: "page.hstack", // --nypl-space-l
+    gridColumn: { base: "1", md: "1 / span 2" },
+    height: "100%",
+    marginBottom: "page.hstack",
   },
 };
 const TemplateContentPrimary = {
   baseStyle: {
-    flex: "1 1",
-    marginBottom: "page.hstack", // --nypl-space-l
+    marginBottom: "page.hstack",
   },
   variants: {
     left: {
-      display: "flex",
-      flexFlow: "column nowrap",
-      order: "1",
+      gridColumn: { base: "1", md: "2" },
       marginEnd: { md: 0 },
       minWidth: { md: 0 },
     },
@@ -83,21 +71,21 @@ const TemplateContentPrimary = {
 };
 const TemplateContentSidebar = {
   baseStyle: {
-    flex: { base: "0 0 auto", md: "0 0 255px" },
-    order: { md: "page.vstack" },
-    marginBottom: "page.hstack", // --nypl-space-l
+    marginBottom: "page.hstack",
   },
   variants: {
     left: {
+      gridColumn: "1",
       marginEnd: {
         base: 0,
-        md: "page.vstack", // --nypl-space-xl
+        md: "page.vstack",
       },
     },
     right: {
+      gridColumn: { base: "1", md: "2" },
       marginStart: {
         base: 0,
-        md: "page.vstack", // --nypl-space-xl
+        md: "page.vstack",
       },
     },
   },
@@ -105,7 +93,6 @@ const TemplateContentSidebar = {
 
 export default {
   Template,
-  TemplateHeader,
   TemplateBreakout,
   TemplateContent,
   TemplateContentTop,